requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.3
decorator==5.2.1
Django==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
jwt==1.3.1
moviepy==2.1.2
numpy==2.2.5
opencv-python==4.11.0.86
pillow==10.4.0
proglog==0.1.12
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2